Must not assume that external functions are pure
It is nowhere said in the MLS that external functions have to be pure and thus can be assigned to a (constant) parameter. Assume that the external function is impure (e.g., a random number generator) then the assignment to a parameter would be invalid.

Since the goal of these models is to test the external "C" support according to MLS 12.9 the parameter keyword should be removed here.

@HansOlsson @sjoelund @perost FYI